+++ src/server/main_server.cpp
@@ -4,7 +4,7 @@
 #include "udpserver.hpp"
 #include "numbersgetter.hpp"
 
-int main() //int argc, char * argv[])
+int main(int argc, char * argv[])
 {
     TCPServer server;
     //char ip[16] = "127.0.0.1";
@@ -12,6 +12,5 @@ int main() //int argc, char * argv[])
     server.setIP(ip);
     server.setPort(49160);
     server.setDataConverter(new NumbersGetter);
-    std::cout << "Server started!" << '\n';
     server.run();
 }
